# example_blog/internal/handler/post_handler_gen.go
// Code generated by apigen. DO NOT EDIT.
package handler

import (
	"net/http"

	"github.com/example/blog/internal/model"
	"github.com/example/blog/internal/service"
	"github.com/gin-gonic/gin"
)

// PostHandler wires HTTP routes to the PostService.
type PostHandler struct {
	svc service.PostService
}

// NewPostHandler constructs a PostHandler.
func NewPostHandler(svc service.PostService) *PostHandler {
	return &PostHandler{svc: svc}
}

// RegisterRoutes registers all Post routes on the given router group.
func (h *PostHandler) RegisterRoutes(r *gin.RouterGroup) {
	r.GET("/posts", h.ListPosts)
	r.GET("/posts/:id", h.GetPost)
	r.POST("/posts", h.CreatePost)
	r.PUT("/posts/:id", h.UpdatePost)
	r.DELETE("/posts/:id", h.DeletePost)
}

// ListPosts handles GET /posts.
func (h *PostHandler) ListPosts(c *gin.Context) {
	var req model.ListPostsRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	result, err := h.svc.ListPosts(c.Request.Context(), req)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, result)
}

// GetPost handles GET /posts/:id.
func (h *PostHandler) GetPost(c *gin.Context) {
	id := c.Param("id")
	result, err := h.svc.GetPost(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, result)
}

// CreatePost handles POST /posts.
func (h *PostHandler) CreatePost(c *gin.Context) {
	var req model.CreatePostRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	result, err := h.svc.CreatePost(c.Request.Context(), req)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, result)
}

// UpdatePost handles PUT /posts/:id.
func (h *PostHandler) UpdatePost(c *gin.Context) {
	var req model.UpdatePostRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	id := c.Param("id")
	result, err := h.svc.UpdatePost(c.Request.Context(), req, id)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, result)
}

// DeletePost handles DELETE /posts/:id.
func (h *PostHandler) DeletePost(c *gin.Context) {
	id := c.Param("id")
	result, err := h.svc.DeletePost(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, result)
}